tests/requirements-test.txt
# Test requirements
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-mock==3.12.0
faker==20.1.0
